mirai_analyser/requirements.txt
pyelftools==0.27
capstone==5.0.6
pefile==2021.9.3
requests==2.26.0
graphviz